refactor(core): extract constants in a file.
